Fix bug in translate_return().
The previous code was trying to hard to be efficient when restoring the
caller's state. I tried to walk back over THAT, THIS, ARG and LCL and
then land on the return address at the end, but if a function has zero
arguments, then the return address will already have been overwritten
with the return value.

With this change, the VM translator is correctly translating
NestedCall, but is still failing for FibonacciElement.