Implement the 'expressionless' syntax analyser.
This analyser is capable of analysing a stream of tokens from the
tokeniser, converting into a sort of abstract syntax tree, and producing
an XML representation of the tree.

It deliberately does not attempt to compile the full 'expression' part
of the grammar (it only handles expressions that are single
identifiers).