Continue work on the compiler for project 11.
This feels like it's very close to being complete. I am unable to
generate VM code for all three source files in the project 11 'Square'
program, and at least on a superficial read, it seems to be sensible.

Next up there will be some debugging to do but this is in a good place.